+++ monkey/monkey_island/cc/ui/src/components/map/preview-pane/PreviewPane.js
@@ -87,7 +87,7 @@ class PreviewPaneComponent extends AuthComponent {
               Download Island Server Log
             </th>
             <td>
-              <IslandLogDownloadButton url={'/api/log/island/download'} />
+              <IslandLogDownloadButton url={'/api/island/log'} />
             </td>
           </tr>
         }